+++ app/src/main/java/com/twyst/app/android/adapters/MenuExpandableAdapter.java
@@ -291,7 +291,7 @@ public class MenuExpandableAdapter extends ExpandableRecyclerAdapter<MenuParentV
                 subOptionSetListNew.add(subOptionSet);
                 subOptionNew.setSubOptionSetList(subOptionSetListNew);
                 option.getSubOptionsList().set(currentIndex, subOptionNew);
-                float itemCostNew = cartItem.getItemCost() + subOptionSet.getSubOptionCost();
+                double itemCostNew = cartItem.getItemCost() + subOptionSet.getSubOptionCost();
                 cartItem.setItemCost(itemCostNew);
                 if ((currentIndex + 1) < option.getSubOptionsList().size()) {
                     showDialogSubOptions(cartItem, currentIndex + 1);
@@ -361,7 +361,7 @@ public class MenuExpandableAdapter extends ExpandableRecyclerAdapter<MenuParentV
                 for (int i = 0; i < menuAddonsAdapter.getSelectedPositions().size(); i++) {
                     AddonSet addonSet = addonsNew.getAddonSetList().get(menuAddonsAdapter.getSelectedPositions().get(i));
                     addonSetListNew.add(addonSet);
-                    float itemCostNew = cartItem.getItemCost() + addonSet.getAddonCost();
+                    double itemCostNew = cartItem.getItemCost() + addonSet.getAddonCost();
                     cartItem.setItemCost(itemCostNew);
                 }
                 addonsNew.setAddonSetList(addonSetListNew);
